requirements.txt
SQLAlchemy==2.0.38 
fastapi[standard]==0.115.8
python-dotenv==1.0.1
psycopg2-binary==2.9.10 